remove custom useDisplayImage hook to separate file and import it
